feat: Configure sqlfmt and completions
This commit adds sqlfmt completions to bash and configures sqlfmt as the formatter for SQL files in Helix. It also moves the SQL language configuration in Helix to be after the YAML configuration. Finally, it updates the `mods-gd` and `mods-gd-continue` functions in Nushell config to remove the unnecessary restriction about explanations.